move wallet logic to WalletLayout from App
